change of mind: only creator of to do is able to delete it; check done

+++ website_resources/logic/back_end/website_pages/pages/dependencies/to_dos/delete_to_do.php
@@ -3,7 +3,6 @@ $initial_load = false;
 require $_SERVER['DOCUMENT_ROOT'] . "/website_resources/logic/back_end/core/global_requirements.php";
 require $_SERVER['DOCUMENT_ROOT'] . "/website_resources/logic/back_end/core/database_connect.php";
 
-$user_id = $_SESSION['user_id'];
 
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     $input = json_decode(file_get_contents('php://input'), true);
@@ -14,12 +13,26 @@ else {
 }
 
 if (empty($to_do_id)) {
-    echo json_encode(['status' => 'error', 'message' => 'Group ID is required.']);
+    echo json_encode(['status' => 'error', 'message' => 'To do ID is required.']);
     exit;
 }
 
-//NEED_TO_DO: need to check: if personal to do - if was written bby user; if group to do - if user is a member of the group
+//NEED_TO_DO DONE: need to check: if personal to do - if was written bby user; if group to do - if user is a member of the group
 // check if personal or group to do in the back end
+
+// - change of mind: only creator of to do is able to delete it
+
+
+$stmt = $conn->prepare("SELECT creator_user_id FROM to_dos WHERE id = ?");
+$stmt->bind_param("i", $to_do_id);
+$stmt->execute();
+$result = $stmt->get_result();
+$row = $result->fetch_assoc();
+ 
+if($row["creator_user_id"] != $_SESSION["user_id"]) {   
+    echo json_encode(["status" => "error", "message" => "User is not allowed to delete the to do."]);
+    exit;
+}
 
 
 $stmt = $conn->prepare("DELETE FROM to_dos WHERE to_do_id = ?");
